Transition to object-oriented approach.
