--- src/main/kotlin/com/example/plugins/JsonData.kt
@@ -1,3 +1,0 @@
-package com.example.plugins
-@kotlinx.serialization.Serializable
-data class JsonData (val t1: Float, val t2:Float, val t3: Float, val p1: Float)
