+++ src/main/kotlin/com/example/plugins/Routing.kt
@@ -5,9 +5,10 @@ import io.ktor.server.application.*
 import io.ktor.server.response.*
 
 fun Application.configureRouting() {
+    val poolInfoController: PoolInfoController = PoolInfoController()
     routing {
         get("/pool-info"){
-            return@get call.respondText(answerServer()!!, status = HttpStatusCode.OK)
+            return@get call.respondText(poolInfoController.answerServer()!!, status = HttpStatusCode.OK)
         }
     }
 }
